+++ workers/ohsome_quality_analyst/indicators/base.py
@@ -2,10 +2,7 @@
 
 import json
 from abc import ABCMeta, abstractmethod
-from dataclasses import asdict, dataclass
-from datetime import datetime, timezone
 from io import StringIO
-from typing import Dict, Literal, Optional
 
 import matplotlib.pyplot as plt
 from geojson import Feature
@@ -15,51 +12,9 @@ from ohsome_quality_analyst.html_templates.template import (
     get_template,
     get_traffic_light,
 )
+from ohsome_quality_analyst.indicators.models import IndicatorMetadata, Result
 from ohsome_quality_analyst.topics.models import BaseTopic as Topic
 from ohsome_quality_analyst.utils.helper import flatten_dict, json_serialize
-
-
-@dataclass
-class Metadata:
-    """Metadata of an indicator as defined in the metadata.yaml file."""
-
-    name: str
-    description: str
-    label_description: Dict
-    result_description: str
-
-
-@dataclass
-class Result:
-    """The result of the Indicator.
-
-    Attributes:
-        timestamp_oqt (datetime): Timestamp of the creation of the indicator
-        timestamp_osm (datetime): Timestamp of the used OSM data
-            (e.g. Latest timestamp of the ohsome API results)
-        label (str): Traffic lights like quality label: `green`, `yellow` or `red`. The
-            value is determined by the result classes
-        value (float): The result value
-        class_ (int): The result class. An integer between 1 and 5. It maps to the
-            result labels. This value is used by the reports to determine an overall
-            result.
-        description (str): The result description.
-        svg (str): Figure of the result as SVG
-    """
-
-    description: str
-    svg: str
-    html: str
-    timestamp_oqt: datetime = datetime.now(timezone.utc)  # UTC datetime object
-    timestamp_osm: Optional[datetime] = None
-    value: Optional[float] = None
-    class_: Optional[Literal[1, 2, 3, 4, 5]] = None
-    figure: Optional[dict] = None
-
-    @property
-    def label(self) -> Literal["green", "yellow", "red", "undefined"]:
-        labels = {1: "red", 2: "yellow", 3: "yellow", 4: "green", 5: "green"}
-        return labels.get(self.class_, "undefined")
 
 
 class BaseIndicator(metaclass=ABCMeta):
@@ -70,16 +25,11 @@ class BaseIndicator(metaclass=ABCMeta):
         topic: Topic,
         feature: Feature,
     ) -> None:
+        self.metadata: IndicatorMetadata = get_metadata(
+            "indicators", type(self).__name__
+        )
         self.topic: Topic = topic
         self.feature: Feature = feature
-        # setattr(object, key, value) could be used instead of relying on from_dict.
-        metadata = get_metadata("indicators", type(self).__name__)
-        self.metadata: Metadata = Metadata(
-            name=metadata["name"],
-            description=metadata["description"],
-            label_description=metadata["label-description"],
-            result_description=metadata["result-description"],
-        )
         self.result: Result = Result(
             description=self.metadata.label_description["undefined"],
             svg=self._get_default_figure(),
@@ -96,7 +46,7 @@ class BaseIndicator(metaclass=ABCMeta):
             flatten (bool): If true flatten the properties.
             include_data (bool): If true include additional data in the properties.
         """
-        result = asdict(self.result)  # only attributes, no properties
+        result = self.result.dict()  # only attributes, no properties
         result["label"] = self.result.label  # label is a property
         result["class"] = result.pop("class_")
         properties = {