+++ workers/ohsome_quality_analyst/reports/base.py
@@ -2,8 +2,7 @@ from __future__ import annotations  # superfluous in Python 3.10
 
 import logging
 from abc import ABCMeta, abstractmethod
-from dataclasses import asdict, dataclass
-from typing import List, Literal, NamedTuple, Tuple
+from typing import List, NamedTuple, Tuple
 
 import numpy as np
 from geojson import Feature
@@ -14,30 +13,8 @@ from ohsome_quality_analyst.html_templates.template import (
     get_traffic_light,
 )
 from ohsome_quality_analyst.indicators.base import BaseIndicator
+from ohsome_quality_analyst.reports.models import ReportMetadata, Result
 from ohsome_quality_analyst.utils.helper import flatten_dict
-
-
-@dataclass
-class Metadata:
-    """Metadata of a report as defined in the metadata.yaml file"""
-
-    name: str
-    description: str
-    label_description: dict
-
-
-@dataclass
-class Result:
-    """The result of the Report."""
-
-    class_: Literal[1, 2, 3, 4, 5] | None = None
-    description: str = ""
-    html: str = ""
-
-    @property
-    def label(self) -> Literal["green", "yellow", "red", "undefined"]:
-        labels = {1: "red", 2: "yellow", 3: "yellow", 4: "green", 5: "green"}
-        return labels.get(self.class_, "undefined")
 
 
 class IndicatorTopic(NamedTuple):
@@ -53,18 +30,13 @@ class BaseReport(metaclass=ABCMeta):
         blocking_red: bool = False,
         blocking_undefined: bool = False,
     ):
+        self.metadata: ReportMetadata = get_metadata("reports", type(self).__name__)
         self.feature = feature
 
         self.indicators: List[BaseIndicator] = []
-        metadata = get_metadata("reports", type(self).__name__)
         self.indicator_topic = indicator_topic  # Defines indicator+topic combinations
         self.blocking_undefined = blocking_undefined
         self.blocking_red = blocking_red
-        self.metadata: Metadata = Metadata(
-            name=metadata["name"],
-            description=metadata["description"],
-            label_description=metadata["label-description"],
-        )
         # Results will be written during the lifecycle of the report object (combine())
         self.result = Result()
 
@@ -74,13 +46,13 @@ class BaseReport(metaclass=ABCMeta):
         The properties of the Feature contains the attributes of all indicators.
         The geometry (and properties) of the input GeoJSON object is preserved.
         """
-        result = asdict(self.result)  # only attributes, no properties
+        result = self.result.dict()  # only attributes, no properties
         result["label"] = self.result.label  # label is a property
         if result["class_"] is not None:
             result["class_"] = self.result.class_
         properties = {
             "report": {
-                "metadata": asdict(self.metadata),
+                "metadata": self.metadata.dict(),
                 "result": result,
             },
             "indicators": [],