+++ workers/tests/unittests/test_definitions.py
@@ -3,19 +3,15 @@ import unittest
 import pytest
 
 from ohsome_quality_analyst import definitions
+from ohsome_quality_analyst.indicators.models import (
+    IndicatorMetadata as IndicatorMetadata,
+)
+from ohsome_quality_analyst.reports.base import ReportMetadata as ReportMetadata
 from ohsome_quality_analyst.topics.models import TopicDefinition
 from ohsome_quality_analyst.utils.exceptions import RasterDatasetUndefinedError
 
 
 class TestDefinitions(unittest.TestCase):
-    def test_load_metadata(self):
-        metadata = definitions.load_metadata("indicators")
-        self.assertIsInstance(metadata, dict)
-
-    def test_get_metadata(self):
-        metadata = definitions.get_metadata("indicators", "Minimal")
-        self.assertIsInstance(metadata, dict)
-
     def test_get_indicator_names(self):
         names = definitions.get_indicator_names()
         self.assertIsInstance(names, list)
@@ -97,3 +93,43 @@ def test_get_topic_definitions_not_found_error():
         definitions.get_topic_definition("foo")
     with pytest.raises(KeyError):
         definitions.get_topic_definition(None)
+
+
+def test_load_metadata_indicator():
+    metadata = definitions.load_metadata("indicators")
+    assert isinstance(metadata, dict)
+    for v in metadata.values():
+        assert isinstance(v, IndicatorMetadata)
+
+
+def test_load_metadata_report():
+    metadata = definitions.load_metadata("reports")
+    assert isinstance(metadata, dict)
+    for v in metadata.values():
+        assert isinstance(v, ReportMetadata)
+
+
+def test_load_metadata_wrong_module():
+    with pytest.raises(AssertionError):
+        definitions.load_metadata("foo")
+    with pytest.raises(AssertionError):
+        definitions.load_metadata("")
+
+
+def test_get_metadata_indicator():
+    metadata = definitions.get_metadata("indicators", "Minimal")
+    assert isinstance(metadata, IndicatorMetadata)
+
+
+def test_get_metadata_report():
+    metadata = definitions.get_metadata("reports", "Minimal")
+    assert isinstance(metadata, ReportMetadata)
+    with pytest.raises(KeyError):
+        definitions.get_metadata("reports", "")
+
+
+def test_get_metadata_wrong_class():
+    with pytest.raises(KeyError):
+        definitions.get_metadata("indicators", "foo")
+    with pytest.raises(KeyError):
+        definitions.get_metadata("indicators", "")